+++ gulp-helpers.js
@@ -1,6 +1,7 @@
 const fs = require('fs');
 const prompts = require('prompts');
 const path = require('path');
+const frontmatter = require('front-matter');
 
 const gulp = require('gulp');
 const clean = require('gulp-clean');
@@ -114,11 +115,36 @@ const patternStates = [
   'complete'
 ];
 
+const filterPatternState = (file, state, callback) => {
+  const filedir = path.parse(file.path).dir;
+  const filename = path.parse(file.path).name;
+  const markdownFile = `${filedir}/${filename}.md`;
+
+  if (fs.existsSync(markdownFile)) {
+    fs.readFile(markdownFile, 'utf8', (err, data) => {
+      if (err) {
+        callback(err);
+      }
+
+      const attributes = frontmatter(data).attributes;
+
+      if (attributes.state !== state) {
+        file = null;
+      }
+
+      callback(null, file);
+    });
+  } else {
+    callback(null);
+  }
+};
+
 module.exports = {
   cleanDir,
   checkForComponents,
   checkForSymlinks,
   importScripts,
   importLess,
-  patternStates
+  patternStates,
+  filterPatternState
 }
